+++ victron_ble/devices/dc_energy_meter.py
@@ -1,30 +1,34 @@
 from enum import Enum
 from typing import Optional
 
-from construct import BitStruct, Flag, Int8sl, Int16sl, Int16ul, Int24sl, Struct
-
-from devices.base import Device, DeviceData, kelvin_to_celsius
-from devices.battery_monitor import AuxMode
+from victron_ble.devices.base import (
+    AlarmReason,
+    BitReader,
+    Device,
+    DeviceData,
+    kelvin_to_celsius,
+)
+from victron_ble.devices.battery_monitor import AuxMode
 
 
 class MeterType(Enum):
-    SOLAR_CHARGER = -7
-    WIND_CHARGER = -6
-    SHAFT_GENERATOR = -5
-    ALTERNATOR = -4
-    FUEL_CELL = -3
-    WATER_GENERATOR = -2
-    DC_DC_CHARGER = -1
-    AC_CHARGER = 1
-    GENERIC_SOURCE = 2
-    GENERIC_LOAD = 3
-    ELECTRIC_DRIVE = 4
-    FRIDGE = 5
-    WATER_PUMP = 6
-    BILGE_PUMP = 7
-    DC_SYSTEM = 8
-    INVERTER = 9
-    WATER_HEATER = 10
+    SOLAR_CHARGER = -9
+    WIND_CHARGER = -8
+    SHAFT_GENERATOR = -7
+    ALTERNATOR = -6
+    FUEL_CELL = -5
+    WATER_GENERATOR = -4
+    DC_DC_CHARGER = -3
+    AC_CHARGER = -2
+    GENERIC_SOURCE = -1
+    GENERIC_LOAD = 1
+    ELECTRIC_DRIVE = 2
+    FRIDGE = 3
+    WATER_PUMP = 4
+    BILGE_PUMP = 5
+    DC_SYSTEM = 6
+    INVERTER = 7
+    WATER_HEATER = 8
 
 
 class DcEnergyMeterData(DeviceData):
@@ -46,41 +50,11 @@ class DcEnergyMeterData(DeviceData):
         """
         return self._data["voltage"]
 
-    def get_low_voltage_alarm(self) -> bool:
+    def get_alarm(self) -> Optional[AlarmReason]:
         """
-        Return a boolean indicating if the low voltage alarm is active
+        Return an enum indicating the current alarm reason or None otherwise
         """
-        return self._data["alarm"]["low_voltage"]
-
-    def get_high_voltage_alarm(self) -> bool:
-        """
-        Return a boolean indicating if the high voltage alarm is active
-        """
-        return self._data["alarm"]["high_voltage"]
-
-    def get_low_starter_battery_voltage_alarm(self) -> bool:
-        """
-        Return a boolean indicating if the low starter battery voltage alarm is active
-        """
-        return self._data["alarm"]["low_starter_voltage"]
-
-    def get_high_starter_battery_voltage_alarm(self) -> bool:
-        """
-        Return a boolean indicating if the high starter battery voltage alarm is active
-        """
-        return self._data["alarm"]["high_starter_voltage"]
-
-    def get_low_temperature_alarm(self) -> bool:
-        """
-        Return a boolean indicating if the low temperature alarm is active
-        """
-        return self._data["alarm"]["low_temperature"]
-
-    def get_high_temperature_alarm(self) -> bool:
-        """
-        Return a boolean indicating if the high temperature alarm is active
-        """
-        return self._data["alarm"]["high_temperature"]
+        return AlarmReason(self._data["alarm"]) if self._data["alarm"] > 0 else None
 
     def get_aux_mode(self) -> AuxMode:
         """
@@ -105,62 +79,41 @@ class DcEnergyMeterData(DeviceData):
 
 
 class DcEnergyMeter(Device):
+    data_type = DcEnergyMeterData
 
-    PACKET = Struct(
-        "meter_type" / Int16sl,
+    def parse_decrypted(self, decrypted: bytes) -> dict:
+        reader = BitReader(decrypted)
+
+        meter_type = reader.read_signed_int(16)
         # Voltage reading in 10mV increments
-        "voltage" / Int16ul,
-        "alarm"
-        / BitStruct(
-            "mid_voltage" / Flag,
-            "high_temperature" / Flag,
-            "low_temperature" / Flag,
-            "high_starter_voltage" / Flag,
-            "low_starter_voltage" / Flag,
-            "low_soc" / Flag,
-            "high_voltage" / Flag,
-            "low_voltage" / Flag,
-        ),
-        # Unknown byte
-        "uk_1b" / Int8sl,
+        voltage = reader.read_signed_int(16)
+        # Alarm reason
+        alarm = reader.read_unsigned_int(16)
         # Value of the auxillary input
-        "aux" / Int16ul,
-        # The upper 22 bits indicate the current in milliamps
-        # The lower 2 bits identify the aux input mode:
+        aux = reader.read_unsigned_int(16)
+        # The aux input mode:
         #   0 = Starter battery voltage
-        #   1 = Midpoint voltage
         #   2 = Temperature
         #   3 = Disabled
-        "current" / Int24sl,
-    )
-
-    def parse(self, data: bytes) -> DcEnergyMeterData:
-        decrypted = self.decrypt(data)
-        pkt = self.PACKET.parse(decrypted)
-
-        aux_mode = AuxMode(pkt.current & 0b11)
+        aux_mode = reader.read_unsigned_int(2)
+        # The current in milliamps
+        current = reader.read_signed_int(22)
 
         parsed = {
-            "meter_type": MeterType(pkt.meter_type),
-            "aux_mode": aux_mode,
-            "current": (pkt.current >> 2) / 1000,
-            "voltage": pkt.voltage / 100,
-            "alarm": {
-                "low_voltage": pkt.alarm.low_voltage,
-                "high_voltage": pkt.alarm.high_voltage,
-                "low_starter_voltage": pkt.alarm.low_starter_voltage,
-                "high_starter_voltage": pkt.alarm.high_starter_voltage,
-                "low_temperature": pkt.alarm.low_temperature,
-                "high_temperature": pkt.alarm.high_temperature,
-            },
+            "meter_type": MeterType(meter_type),
+            "aux_mode": AuxMode(aux_mode),
+            "current": current / 1000 if current != 0x3FFFFF else None,
+            "voltage": voltage / 100 if voltage != 0x7FFF else None,
+            "alarm": alarm,
         }
 
-        if aux_mode == AuxMode.STARTER_VOLTAGE:
+        if aux_mode == AuxMode.STARTER_VOLTAGE.value:
             # Starter voltage is treated as signed
-            parsed["starter_voltage"] = (
-                Int16sl.parse((pkt.aux).to_bytes(2, "little")) / 100
-            )
-        elif aux_mode == AuxMode.TEMPERATURE:
-            parsed["temperature_kelvin"] = pkt.aux / 100
+            parsed["starter_voltage"] = BitReader.to_signed_int(aux, 16) / 100
+        elif aux_mode == AuxMode.TEMPERATURE.value:
+            if aux == 0xFFFF:
+                parsed["temperature_kelvin"] = None
+            else:
+                parsed["temperature_kelvin"] = aux / 100
 
-        return DcEnergyMeterData(self.get_model_id(data), parsed)
+        return parsed