+++ HISTORY.md
@@ -1,6 +1,20 @@
 Changelog
 =========
 
+0.6.0 (2023)
+------------------
+- Add VEBus support
+- Improve readme 
+
+0.5.1 (2023)
+------------------
+- Fix DC/DC output voltage when off
+
+0.5.0 (2023)
+------------------
+- Add DC/DC charger support
+- [breaking] Change format of alarm getters
+
 0.4.0 (2022)
 ------------------
 - Add solar charger support